rename tccpencryption to encryptionensurer resource
