+++ service/controller/resource/encryptionensurer/delete.go
@@ -1,4 +1,4 @@
-package tccpencryption
+package encryptionensurer
 
 import (
 	"context"
